MAINT: split _clean_text_signature out to _utils.
All other functions defined in numpydoc.py module depend on sphinx
and therefore can be skipped for testing purposes when sphinx is
not installed.

importorskips the numpydoc tests but maintains the testing of
_clean_text_signature in new test_utils.py module.
