+++ numpydoc/tests/test_numpydoc.py
@@ -1,15 +1,17 @@
+import pytest
+pytest.importorskip("sphinx")
+pytest.importorskip("docutils")
+
 from collections import defaultdict
 from copy import deepcopy
 from io import StringIO
 from pathlib import PosixPath
 
-import pytest
 from docutils import nodes
 from sphinx.ext.autodoc import ALL
 from sphinx.util import logging
 
 from numpydoc.numpydoc import (
-    _clean_text_signature,
     clean_backrefs,
     mangle_docstrings,
     update_config,
@@ -109,40 +111,6 @@ A top section before
     mangle_docstrings(app, "class", "pathlib.PosixPath", PosixPath, {}, lines)
     lines = [x.strip() for x in lines]
     assert "samefile" not in lines
-
-
-def test_clean_text_signature():
-    assert _clean_text_signature(None) is None
-    assert _clean_text_signature("func($self)") == "func()"
-    assert (
-        _clean_text_signature("func($self, *args, **kwargs)") == "func(*args, **kwargs)"
-    )
-    assert _clean_text_signature("($self)") == "()"
-    assert _clean_text_signature("()") == "()"
-    assert _clean_text_signature("func()") == "func()"
-    assert (
-        _clean_text_signature("func($self, /, *args, **kwargs)")
-        == "func(*args, **kwargs)"
-    )
-    assert (
-        _clean_text_signature("func($self, other, /, *args, **kwargs)")
-        == "func(other, *args, **kwargs)"
-    )
-    assert _clean_text_signature("($module)") == "()"
-    assert _clean_text_signature("func($type)") == "func()"
-    assert (
-        _clean_text_signature('func($self, foo="hello world")')
-        == 'func(foo="hello world")'
-    )
-    assert (
-        _clean_text_signature("func($self, foo='hello world')")
-        == "func(foo='hello world')"
-    )
-    assert _clean_text_signature('func(foo="hello world")') == 'func(foo="hello world")'
-    assert _clean_text_signature('func(foo="$self")') == 'func(foo="$self")'
-    assert _clean_text_signature('func($self, foo="$self")') == 'func(foo="$self")'
-    assert _clean_text_signature("func(self, other)") == "func(self, other)"
-    assert _clean_text_signature("func($self, *args)") == "func(*args)"
 
 
 @pytest.fixture
